chore: add component GeometryCanvas
