+++ src/modules/geometry/components/geometry-canvas/GeometryCanvasComponent.tsx
