+++ src/modules/geometry/components/geometry-canvas/index.ts
@@ -1,0 +1,11 @@
+import GeometryCanvasComponent from "src/modules/geometry/components/geometry-canvas/GeometryCanvasComponent";
+import GeometryCanvasConfig from "src/modules/geometry/components/geometry-canvas/GeometryCanvasConfig";
+import GeometryCanvasLoader from "src/modules/geometry/components/geometry-canvas/GeometryCanvasLoader";
+import GeometryCanvasRoot from "src/modules/geometry/components/geometry-canvas/GeometryCanvasRoot";
+
+export const GeometryCanvas = {
+  Component: GeometryCanvasComponent,
+  Config: GeometryCanvasConfig,
+  Loader: GeometryCanvasLoader,
+  Root: GeometryCanvasRoot,
+};
