--- src/shared/ui/geometry-canvas/index.ts
@@ -1,11 +1,0 @@
-import GeometryCanvasComponent from "src/shared/ui/geometry-canvas/GeometryCanvasComponent";
-import GeometryCanvasConfig from "src/shared/ui/geometry-canvas/GeometryCanvasConfig";
-import GeometryCanvasLoader from "src/shared/ui/geometry-canvas/GeometryCanvasLoader";
-import GeometryCanvasRoot from "src/shared/ui/geometry-canvas/GeometryCanvasRoot";
-
-export const GeometryCanvas = {
-  Component: GeometryCanvasComponent,
-  Config: GeometryCanvasConfig,
-  Loader: GeometryCanvasLoader,
-  Root: GeometryCanvasRoot,
-};
